Feature#main#hardening (#36)
* harden(#11,#3,#2): DSN redaction, RAII temp-dir cleanup, 128-bit cache key

Three independent hardening fixes developed together on this branch:

#11 Credential redaction in logs: new kokedb_common::redact::redact_dsn masks
the password in any `scheme://user:pass@host` DSN; the catalog-sync log sites
now log the redacted form instead of the raw DSN (which leaked the meta-store
password into logs).

#3 RAII temp-dir cleanup: new kokedb_common::file::TempPath removes its
directory on Drop unless disarmed. The sync runner arms a guard over the
in-progress snapshot dir (disarmed only after the schema is committed) plus the
flat/delta scratch dirs, and shadow validation guards its per-run temp base —
so a failure or early return no longer leaks partial snapshots to disk.

#2 128-bit cache key: a 64-bit plan hash has a non-negligible birthday
collision probability at scale, and a collision would serve one query's cached
rows for a different query — a silent correctness bug. get_plan_hash now uses
xxh3_128; u64 -> u128 is propagated through the whole cache-key chain (query,
plan resolver, catalog manager, meta store, task manager, MySQL server, the
foyer hybrid cache, singleflight). The meta columns that persist these keys
(sql_stats.sql_hash, table_stats.cache_key_list) widen from VARCHAR(20) to
VARCHAR(40) with idempotent ALTER ... TYPE migrations for existing deployments.


* harden(#1): reuse the shared meta pool on hot paths

`PostgreSQLMetaCatalogProviderList::new()` opens a brand-new PgPool every
call. It was being called on the request path (per query, in save_sql_history)
and on the scheduler path (per catalog, per tick, in tick_refresh_due /
reevaluate_catalog / sweep_audits) — connection churn that grows with QPS and
catalog count.

Thread the process-wide shared meta handle (SharedServices.catalog_list,
exposed via a new SharedServices::meta() accessor; dynamic_catalog_list in the
catalog manager) into all of these instead of constructing a fresh pool:
  - save_sql_history now takes &meta; the MySQL server carries it per
    connection (CoreContex.meta) from shared.meta().
  - reevaluate_catalog / tick_refresh_due / sweep_audits now take an
    Arc<meta>; the scheduler tick closure clones the one shared handle.


---------

+++ src/plan/src/lib.rs
@@ -55,7 +55,7 @@ pub async fn resolve_and_execute_plan(
     ctx: &SessionContext,
     config: Arc<PlanConfig>,
     plan: spec::Plan,
-    key: u64,
+    key: u128,
 ) -> PlanResult<Arc<dyn ExecutionPlan>> {
     let resolver = PlanResolver::new(ctx, config, key);
     let NamedPlan { plan, fields } = resolver.resolve_named_plan(plan).await?;
